+++ readalign/__init__.py
@@ -6,7 +6,7 @@ from .silence import energy_frames, held, speech_level, speech_threshold
 from .weighting import EnglishSyllableWeighting, EvenWeighting, SpeechWeighting
 from .words import RecognizedWord, WordMatch, WordSpan, fold, normalize, printed_parts, spoken
 
-__version__ = "0.13.1"
+__version__ = "0.14.0"
 __all__ = [
     "EnglishSyllableWeighting",
     "Equivalent",